+++ common/src/main/kotlin/terrasect/definition/Region.kt
@@ -1,20 +1,20 @@
 package terrasect.definition
 
 class Region(
-  val name: String,
-  val budget: Double,
-  val children: Set<Region>,
-  val strategy: Strategy? = null,
-  val climate: ClimateSettings? = null,
-  val height: HeightConstraints? = null,
-  val noise: NoiseConstraints? = null,
-  val biomes: SelectionRules? = null,
-  val structures: SelectionRules? = null,
-  val mobs: SelectionRules? = null,
+    val name: String,
+    val budget: Long,
+    val children: Set<Region>,
+    val strategy: Strategy? = null,
+    val climate: ClimateSettings? = null,
+    val height: HeightConstraints? = null,
+    val noise: NoiseConstraints? = null,
+    val biomes: SelectionRules? = null,
+    val structures: SelectionRules? = null,
+    val mobs: SelectionRules? = null,
 ) {
   val hasChildren = children.isNotEmpty()
 
   companion object {
-    fun empty(name: String): Region = Region(name, 10000.0, emptySet())
+    fun empty(name: String): Region = Region(name, 10000, emptySet())
   }
 }
