+++ common/src/main/kotlin/terrasect/definition/RegionDefinition.kt
@@ -6,7 +6,7 @@ open class RegionDefinition(
     val registry: RegionRegistry,
     val name: String,
     val originAnchor: Boolean = false,
-    var budget: Double,
+    var budget: Long,
     val strategy: StrategySettings,
     var adjacentTo: Set<String>? = null,
     var parent: String? = null,
@@ -21,7 +21,7 @@ open class RegionDefinition(
 
 open class RegionBuilder(val registry: RegionRegistry, var name: String) {
   var originAnchor = false
-  var budget: Double? = null
+  var budget: Long? = null
   var strategy: StrategySettings? = null
   val adjacentToLazy = lazy { setOf<String>() }
   val adjacentTo by adjacentToLazy
@@ -41,7 +41,7 @@ open class RegionBuilder(val registry: RegionRegistry, var name: String) {
   val mobsLazyBuilder = lazy { SelectionRules.builder() }
   val mobsBuilder by mobsLazyBuilder
 
-  fun area(budget: Double) = apply { this.budget = budget * budget }
+  fun area(budget: Long) = apply { this.budget = budget * budget }
 
   fun originAnchor() = apply { originAnchor = true }
 
@@ -117,13 +117,13 @@ open class RegionBuilder(val registry: RegionRegistry, var name: String) {
   }
 
   fun build(children: Set<Region>? = null): RegionDefinition {
-    val budget = children?.sumOf { it.budget }?.let { max(this.budget ?: 0.0, it) }
+    val budget = children?.sumOf { it.budget }?.let { max(this.budget ?: 0, it) }
 
     return RegionDefinition(
         registry = this.registry,
         name = this.name,
         originAnchor = this.originAnchor,
-        budget = budget ?: (this.budget ?: 0.0),
+        budget = budget ?: (this.budget ?: 0),
         strategy = this.strategy ?: Strategy.voronoi(),
         adjacentTo = if (this.adjacentToLazy.isInitialized()) adjacentTo else null,
         parent = this.parent,
