+++ common/src/main/kotlin/terrasect/generation/Traverse.kt
@@ -1,10 +1,10 @@
 package terrasect.generation
 
+import java.nio.ByteBuffer
 import terrasect.cache.Cache
 import terrasect.definition.Region
 import terrasect.definition.Strategy
 import terrasect.sdf.SdfCompose
-import java.nio.ByteBuffer
 
 open class Traverse(
     val seed: Long,
@@ -13,14 +13,14 @@ open class Traverse(
   val iterator: ThreadLocal<TraversalStep>
     get() = ThreadLocal.withInitial { TraversalStep(this) }
 
-  fun iterate(x: Double, z: Double, cache: Cache? = null): TraversalStep {
+  fun iterate(x: Int, z: Int, cache: Cache? = null): TraversalStep {
     val step = this.iterator.get()
     step.reset(x, z, cache)
 
     return step
   }
 
-  fun traverse(x: Double, z: Double, cache: Cache? = null): TraversalStep {
+  fun traverse(x: Int, z: Int, cache: Cache? = null): TraversalStep {
     var step = this.iterate(x, z, cache)
 
     while (step.region.hasChildren) {
@@ -37,11 +37,11 @@ class TraversalStep(val traverse: Traverse) {
 
   var cache: Cache? = null
   var region: Region = traverse.root
-  var x: Double = 0.0
-  var z: Double = 0.0
-  var distance: Double = Double.NEGATIVE_INFINITY
+  var x: Int = 0
+  var z: Int = 0
+  var distance: Float = Float.NEGATIVE_INFINITY
 
-  fun reset(x: Double, z: Double, cache: Cache? = null) {
+  fun reset(x: Int, z: Int, cache: Cache? = null) {
     this.id.clear()
     this.sdf.reset()
 
@@ -49,7 +49,7 @@ class TraversalStep(val traverse: Traverse) {
     this.region = traverse.root
     this.x = x
     this.z = z
-    this.distance = Double.NEGATIVE_INFINITY
+    this.distance = Float.NEGATIVE_INFINITY
   }
 
   fun next(): TraversalStep? {
